Read and report system memory map

+++ kernel/t.h
@@ -9,11 +9,11 @@ typedef	unsigned char		byte;
 
 typedef	signed char		int8;
 typedef	signed short		int16;
-typedef	signed int		int32;
+typedef	signed long		int32;
 typedef	signed long long	int64;
 
 typedef	unsigned int		uint;
 typedef	unsigned char		uint8;
 typedef	unsigned short		uint16;
-typedef	unsigned int		uint32;
+typedef	unsigned long		uint32;
 typedef	unsigned long long	uint64;
